Updated GameTemplate to fit our idea
Updated a lot of the game to start adapting to our final idea better, ending a match works with a button but it seems like starting a new one is not working.
We are unsure if it's just creating a new match or if it's because these players have been in a match before.